Added in hardware connection functionality
When the program starts (or on demand), the testConnection
function is called for each HardwareObject. The hardwareManager
keeps track of what's successful, and will generate appropriate
errors if there is a critical failure.

More work needs to be done with the hardwareFailure signal; it
should have facility for aborting ongoing acquisitions if failure
is critical.