Added FtmwConfig class
Each sub-experiment will have its own configuration data. Data
will only be written to disk if that particular sub-experiment
is enabled in the experiment as a whole. At present, just the
basic functionality is in place, as well as a prototype for
generating a header.

The headers will be aggregated in a QHash object, with values
that are QPair<QVariant,QString> for the value and unit. Hashses
from all data sources will be aggregated, converted to a map
for alphabetizing, and then written to a header file.