Refactor to make parallelization easier. Step 1.
A raw data vector is kept in the FtmwConfig that has all data
stored in contiguous memory so that GPU transfers can be
made most quickly. Parsing of waveform data now places all elements
in a vector instead of a list of fids. The list is created anew
each time data are averaged in. Profiling shows this approach
is equally efficient as the old approach, but I'm still unhappy
because it feels like one extra iteration over memory space is
made than is absolutely necessary (an extra copy or initialization).
Could perhaps be fixed by avoiding QVector class in Fid object, and
instead using a raw C-style array.