Added basic implementation of tracking plots.
This is intended to be an improved version of the tracking plots
from the BBAcq code. The TrackingViewWidget can accomodate
anywhere from 1-9 plots, and the number can be changed from the UI
at any time. Some of that functionality is yet untested, because I
only have one curve right now (and no implementation of moving the
curve to a new plot), so I haven't been able to verify that moving
curves when plot number is changed works gracefully.