Fixed timers, which suddenly stopped working.
Very strange issue. The timers were working before, but should not
have been. Timers need to be created in the thread where they are
used, and I had been stack allocating them at their parent object
creation (which was in the main thread). Not sure why it was
even working before, or why it suddenly stopped, but the issue is
now fixed.