Change communication initialization.
This change is intended to make it easier to switch between
multiple hardware implementations that use different communication
protocols. Previously, GPIB devices required different
initialization code in hardwaremanager.cpp, which defeated
modularity. Now all devices are initialized in the same manner.

Note that the d_threaded protected variable in HardwareObject
is used to indicate whether a device should get its own thread of
execution. All GPIB devices must share the thread of the gpib
controller (currently). In the future, there is the possibility
of associating different devices with different GPIB hubs, but
that is not supported at present.