Change strings in ChirpConfig hdr and chp file
For historical reasons, the term "Valon" was used in a couple
of keys referring to the Tx oscillator. These were removed to
avoid confusion for users who do not use a Valon synth. The
changes are backwards-compatible; the program will still be able
to read old files that contain "Valon".