Bugfixes for motor scan and interoperability
Added a new variable to HardwareObject called
d_enabledForExperiment that allows individual pieces of
hardware to be disabled for a particular experiment. The
main use case, for instance, is to ensure that during a
motor scan, the AWG and FTMW oscilloscopes are not sent
into acquisition mode. In versions of the program compiled
without LIF and/or motor capability, this should have no
effect.