Allow user to control when TWT pulse goes low
In addition, the minimum delay values are now clearly placed
in the settings file on first run, and a note has been added
that allows the user to change them. For instance, my TWT has
a 100 ns lag between when the enable pulse goes low and when
the amp actually switches off, so I want the post chirp delay
to be -100 ns. By editing the settings file, I can tweak that.

However, changing those settings can be dangerous if someone
is careless. It's now theoretically possibe for someone to disable
the protection during the chirp, so use with caution.