Simplify and standardize HW initialization/tests
To try and minimize the amount of work needed to properly
integrate new hardware with blackchirp, common tasks related to
settings, initialization, and connection testing have been moved
into the HardwareObject base class. The initialize and testConnection
functions have been made protected, and the HardwareManager now calls
public wrapper functions bcInitInstrument and bcTestConnection.
Those wrapper functions ensure that testConnection is always called
during initializtion, and that the appropriate connection signals
are emitted.

Connection error messages are now stored in a d_errorString member,
and are retrieved by the base class to send on to other parts of the
program.