Simpler workflow for time data
Time data recording is now streamlined. The HardwareObject class
now has 2 virtual functions: one for time data that is plotted, and
another for time data that is not plotted. These functions just
need to return a list of string variant pairs containing the data.
HardwareObject has a default implementation for each that does nothing,
so classes that do not need time data don't have to deal with it at all.
The timeDataRead signal is now private to HardwareObject and will
be emitted by the base class as appropriate.