Finish new HW initialization/cleanup
Hardware objects are now created before the UI is initialized.
Now hardware settings that affect the UI can be set in hardware
constructors. This eliminates the need for the number of flow
and pulse channels to be set with preprocessor definitions.
Also made some HardwareObject functions have default implementations
to reduce the amount of boilerplate code needed to implement a new
type. Finally, added override and final specifications to hardware
headers to make the inheritance patterns more explicit.