Fix inconsistency with number of frames in FF mode
Tektronix scopes add an extra frame when using a summary
frame for averaging. This extra frame was being added to the
frame count, which was in turn used for calculating the number
of shots (FIDs). This update corrects this probem, and makes the
shot count correct in fast frame mode with a summary frame.