Significant improvements to python module.
More settings for peak finding and fitting are now user-configurable
both by default and for a particular experiment. Instead of bare
class attributes, there is now a BlackChirpExperiment._settings
dict with the user-adjustable parameters. The keys are printed on
startup

Also added a fitting function that constrains all peaks in a chunk
to have the same linewidth (bc_multigauss_fitw), and improved
performance of fitting by vectorizing the model function calls.
Fitting >2000 peaks now takes only 10 or 15 seconds.