Overhaul LIF data treatment; other misc fixes
This does touch on motor and FTMW code, becuase I modified
some of the ZoomPanPlot functionality. I also noticed that the FTPlot
"Export XY" functionality had been inadvertently disabled, so
that has been re-enabled. It would have been better to split that
change out, but it wasn't going to be easy