Remove implicit sharing from Experiment class
Experiments are only passed around the code during
initialization and saving; they do not need to be
implicitly shared. This, in turn, allows them to be passed
by reference during hardware initialization, since that is
done serially anyways. There remains a lot of cleaning up during
some hardware initialization to make the code more elegant,
but it's still functional.