New constructor for SettingsStorage; start using
Added a new constructor that allows a SettingsStorage instance to
look up the current subKey associated with a piece of hardware.
New tests were written, and the new constructor is used in
ChirpTableModel to read AWG limits. The LogHandler is also now
using SettingsStorage to look up the path to saving log data.