Add new BlackchirpPlotCurve class
Accidentally got sidetracked while implementing settings, but this
is going to dramatically streamline all of the plot classes. The
new BlackchirpPlotCurve class will keep track of a curve's data and
settings (including things like which axes it's plotted on). In
addition, the curve itself will support data filtering and exporting
as XY data, so it will eliminate lots of copy/paste code between
classes. Lots of functionality is moving into the ZoomPanPlot
base class.