Starting the major FTMW refactor
Removed implicit sharing from FtmwConfig, created new base
class for FidStorage and one real implementation. Instead of
passing around copies of the entire FtmwConfig, the UI will just
retain a shared_ptr to the FidStorage and any other needed
information at the initialization stage.

For DR scans and LO scans, this will have the advantage of
conserving memory, and should in thoery improve performance
somewhat by minimizing the number of deep copies made.

However, at the moment this is a major work in progess, and the
code is not in a functional state for this commit!